fix: refactor main runner off of index.ts
